Fix typo where intptr_t should have been used instead of long
This was caught on Flutter Windows buildbot: https://build.chromium.org/p/client.flutter/builders/Windows%20Engine/builds/682

+++ public/lib/fxl/files/file.cc
@@ -111,7 +111,7 @@ std::pair<uint8_t*, intptr_t> ReadFileToBytes(const std::string& path) {
     offset += bytes_read;
     bytes_left -= bytes_read;
   }
-  return std::pair<uint8_t*, long>(ptr, file_size);
+  return std::pair<uint8_t*, intptr_t>(ptr, file_size);
 }
 
 bool IsFile(const std::string& path) {
